+++ lib/runners/processor/jshint.rb
@@ -18,7 +18,7 @@ module Runners
     register_config_schema(name: :jshint, schema: Schema.runner_config)
 
     def setup
-      add_warning_if_deprecated_options([:options])
+      add_warning_if_deprecated_options
       yield
     end