+++ lib/runners/processor/rails_best_practices.rb
@@ -43,7 +43,7 @@ module Runners
     }.freeze
 
     def setup
-      add_warning_if_deprecated_options([:options])
+      add_warning_if_deprecated_options
 
       prepare_config
       install_gems default_gem_specs, optionals: OPTIONAL_GEMS, constraints: CONSTRAINTS do